requirements.txt
alembic==1.14.1
algokit==2.5.2
algokit-utils==2.4.0
algopy==0.7.2
algosdk==2.7.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.2
bases==0.3.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cmake==3.31.4
colorama==0.4.6
copier==9.4.1
cryptography==43.0.3
Deprecated==1.2.15
distlib==0.3.9
distro==1.9.0
docstring-parser==0.14.1
dunamai==1.23.0
executing==2.0.1
fastapi==0.115.6
filelock==3.17.0
flake8==7.1.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
funcy==2.0
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jinja2-ansible-filters==1.3.2
jiter==0.8.2
json5==0.10.0
jsondiff==2.2.1
keyring==25.2.1
linkedin-api==2.3.1
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.5.0
msgpack==1.1.0
mslex==1.3.0
multiformats==0.3.1
multiformats-config==0.3.1
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.2
openai==1.61.1
packaging==24.2
pathspec==0.12.1
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
prompt_toolkit==3.0.48
py-algorand-sdk==2.7.0
pyclip==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
pyteal==0.26.1
pytest==8.3.4
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
questionary==1.10.0
requests==2.32.3
rfc3986==1.5.0
scikit-build==0.18.1
scipy==1.15.1
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
tabulate==0.9.0
tqdm==4.67.1
typing-validation==1.2.11.post4
typing_extensions==4.12.2
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.34.0
virtualenv==20.29.1
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
